made modification to store test case
